% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wandb_agent.R
\name{wandb_agent}
\alias{wandb_agent}
\title{wandb_agent}
\usage{
wandb_agent(
  sweep_id = NULL,
  func = NULL,
  entity = NULL,
  project = NULL,
  count = NULL
)
}
\arguments{
\item{sweep_id}{(dict) Sweep ID generated by CLI or sweep API}

\item{func}{(func, optional) A function to call instead of the "program" specifed in the config.}

\item{entity}{(str, optional) W&B Entity}

\item{project}{(str, optional) W&B Project}

\item{count}{(int, optional) the number of trials to run.}
}
\value{

}
\description{
Will run a function or program with configuration parameters specified by server
}
