% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wandb_init.R
\name{wandb_init}
\alias{wandb_init}
\title{wandb_init}
\usage{
wandb_init(
  project = NULL,
  learning_rate = NULL,
  architecture = NULL,
  layer_1 = NULL,
  activation_1 = NULL,
  dropout = NULL,
  layer_2 = NULL,
  activation_2 = NULL,
  optimizer = NULL,
  loss = NULL,
  metric = NULL,
  epoch = NULL,
  batch_size = NULL,
  dataset = NULL,
  config = NULL
)
}
\arguments{
\item{dataset}{}
}
\value{

}
\description{
Starts a new run to track and log to W&B.
}
\details{
wandb_init spawns a new background process to log data to a run, and it also syncs data to wandb_ai by default so you can see live visualizations. Call wandb_init to start a run before logging data with wandb_log.
wandb_init returns a run object, and you can also access the run object with wandb.run.
At the end of your script, we will automatically call wandb_finish to finalize and cleanup the run. However, if you call wandb_init from a child process, you must explicitly call wandb_finish at the end of the child process.
}
