% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wandb_log.R
\name{wandb_log}
\alias{wandb_log}
\title{wandb_log}
\usage{
wandb_log(acc = NULL, loss = NULL, epoch = NULL)
}
\arguments{
\item{acc}{accuracy}

\item{loss}{loss}

\item{epoch}{epoch}
}
\value{

}
\description{
Logs a dictionary of data to the current run's history.
}
