% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wandb_login.R
\name{wandb_login}
\alias{wandb_login}
\title{Logs into wandb}
\usage{
wandb_login(api_key)
}
\arguments{
\item{api_key}{your wandb api key}
}
\value{
a string indicating login success
}
\description{
Logs into wandb
}
