% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wandb_init.R
\name{wandb_init}
\alias{wandb_init}
\title{Title}
\usage{
wandb_init(
  project = NULL,
  learning_rate = NULL,
  architecture = NULL,
  layer_1 = NULL,
  activation_1 = NULL,
  dropout = NULL,
  layer_2 = NULL,
  activation_2 = NULL,
  optimizer = NULL,
  loss = NULL,
  metric = NULL,
  epoch = NULL,
  batch_size = NULL,
  dataset = NULL
)
}
\arguments{
\item{dataset}{}
}
\value{

}
\description{
Title
}
